Changed speaker disconnect behavior
Disconnecting the speaker used to toggle the sound on and off.
Now it just stops the sound.
This prevents the app from making sound when it is running in the
background, but not in use, and you disconnect a speaker.

Bumped the version number to 1.0.1 and the VersionCode to 2.

Also looks like Android studio inserted a bunch of garbage since I
opened it from the android sub-folder and not the project root.